It was just one parameter...
Lots of unnecessary code here to try to fix the problem with "fuzzy" line joining.  Minor-in and minor-out positions are also given as full regions now, with their own name based on their position in the nucleosome.  I'm keeping it around in case it's useful in the future, but right now it's not actually utilized.  Turns out 'lineend = "round"' was all I really needed...